move legacy physics into separate mod
