+++ src/scene/legacy_physics.rs
@@ -4,10 +4,12 @@ use crate::{
     core::{pool::Handle, visitor::prelude::*},
     engine::PhysicsBinder,
     physics3d::{
-        body::RigidBodyContainer,
-        collider::ColliderContainer,
-        desc::{ColliderShapeDesc, PhysicsDesc},
-        joint::JointContainer,
+        legacy::body::RigidBodyContainer,
+        legacy::collider::ColliderContainer,
+        legacy::desc::{ColliderShapeDesc, PhysicsDesc},
+        legacy::joint::JointContainer,
+        legacy::PhysicsWorld,
+        legacy::RigidBodyHandle,
         rapier::{
             dynamics::{JointSet, RigidBodySet},
             geometry::{Collider, ColliderBuilder, ColliderSet},
@@ -17,7 +19,6 @@ use crate::{
             },
             parry::shape::SharedShape,
         },
-        PhysicsWorld, RigidBodyHandle,
     },
     scene::{graph::Graph, node::Node},
     utils::log::{Log, MessageKind},
